+++ src/main/java/com/buuz135/simpleclaims/commands/CommandMessages.java
@@ -55,4 +55,7 @@ public class CommandMessages {
     public static final Message PLAYER_PARTY_CHAT_ACTIVATED = Message.translation("commands.simpleclaims.playerActivatedPartyChat").color(Color.GREEN).bold(true);
     public static final Message PLAYER_PARTY_CHAT_DEACTIVATED = Message.translation("commands.simpleclaims.playerDeactivatedPartyChat").color(Color.GREEN).bold(true);
     public static final Message NO_PERMISSION = Message.translation("commands.errors.simpleclaims.noPermission").color(Color.RED).bold(true);
+
+    public static final Message CHUNK_NOT_ADJACENT = Message.translation("commands.errors.simpleclaims.chunkNotAdjacent").color(Color.RED).bold(true);
+    public static final Message CHUNK_RESERVED_BY_OTHER_PARTY = Message.translation("commands.errors.simpleclaims.chunkReservedByOtherParty").color(Color.RED).bold(true);
 }