+++ src/main/java/com/buuz135/simpleclaims/config/SimpleClaimsConfig.java
@@ -3,7 +3,6 @@ package com.buuz135.simpleclaims.config;
 
 public class SimpleClaimsConfig {
 
-
     private String[] PartyCommandAliases = new String[]{"scp", "sc-party", "party"};
     private int DefaultPartyClaimsAmount = 25;
     private int MaxAddChunkAmount = 100; // The maximum amount of chunks a party can have when using the add-chunk-amount command
@@ -50,6 +49,9 @@ public class SimpleClaimsConfig {
 
     private boolean ForceSimpleClaimsChunkWorldMap = true;
     private boolean CreativeModeBypassProtection = false;
+    private boolean EnableAdjacentChunkRestriction = false;
+    private boolean EnablePerimeterReservation = false;
+    private boolean ShowPerimeterReservationOnTheMap = false;
 
     private String[] BlocksThatIgnoreInteractRestrictions = new String[]{"gravestone"};
 
@@ -231,6 +233,16 @@ public class SimpleClaimsConfig {
 
     public boolean isMigrateOldClaimOverrides() {
         return MigrateOldClaimOverrides;
+    public boolean isEnableAdjacentChunkRestriction() {
+        return EnableAdjacentChunkRestriction;
+    }
+
+    public boolean isEnablePerimeterReservation() {
+        return EnablePerimeterReservation;
+    }
+
+    public boolean isShowPerimeterReservationOnTheMap() {
+        return ShowPerimeterReservationOnTheMap;
     }
 
 }